# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for WeCare Holidays applications",
        "title": "WeCare Holidays API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.wecareholidays.com/support",
            "email": "support@wecareholidays.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/organizations/{id}/logo": {
            "post": {
                "description": "Upload a new logo for an organization and update its record",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizations"
                ],
                "summary": "Upload organization logo",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"6824886e6b180b753cea43e9\"",
                        "description": "Organization ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Logo image file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated organization",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.SwaggerStandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.Organization"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Organization not found",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "413": {
                        "description": "File too large",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/organizations/{id}/status": {
            "put": {
                "description": "Update the status of an organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizations"
                ],
                "summary": "Update organization status",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"6824886e6b180b753cea43e9\"",
                        "description": "Organization ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.SwaggerStandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.Organization"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Address": {
            "description": "Physical address information",
            "type": "object",
            "properties": {
                "city": {
                    "description": "City name",
                    "type": "string",
                    "example": "Siliguri"
                },
                "country": {
                    "description": "Country name",
                    "type": "string",
                    "example": "India"
                },
                "pincode": {
                    "description": "Postal/ZIP code",
                    "type": "string",
                    "example": "734001"
                },
                "state": {
                    "description": "State or province",
                    "type": "string",
                    "example": "West Bengal"
                },
                "street": {
                    "description": "Street address including building/apartment",
                    "type": "string",
                    "example": "123 Main Street, Building A"
                }
            }
        },
...
        "entity.Organization": {
            "description": "Organization entity representing clients, suppliers, or travel agents",
            "type": "object",
            "properties": {
                "address": {
                    "description": "Physical address",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Address"
                        }
                    ]
                },
                "createdAt": {
                    "description": "Creation timestamp",
                    "type": "string"
                },
                "deletedAt": {
                    "description": "Soft delete timestamp",
                    "type": "string"
                },
                "email": {
                    "description": "Primary contact email",
                    "type": "string",
                    "example": "contact@wecareholidays.com"
                },
                "id": {
                    "description": "Unique identifier for the organization",
                    "type": "string",
                    "example": "5f8d0c1b7ea3f0d0f3c8e1b9"
                },
                "logo": {
                    "description": "URL to organization logo",
                    "type": "string",
                    "example": "https://storage.example.com/logos/wecare.png"
                },
                "name": {
                    "description": "Organization name",
                    "type": "string",
                    "example": "WeCare Holidays"
                },
                "phone": {
                    "description": "Contact phone number",
                    "type": "string",
                    "example": "+91 1234567890"
                },
                "slug": {
                    "description": "URL-friendly version of name",
                    "type": "string",
                    "example": "wecare-holidays"
                },
                "status": {
                    "description": "Current status (Pending, Approved, Suspended, Archived)",
                    "type": "string",
                    "example": "Approved"
                },
                "taxIds": {
                    "description": "Tax identification numbers",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['GST123456'",
                        " 'PAN1234567']"
                    ]
                },
                "type": {
                    "description": "Organization type (SUPPLIER, TRAVEL_AGENT, PLATFORM)",
                    "type": "string",
                    "example": "SUPPLIER"
                },
                "updatedAt": {
                    "description": "Last update timestamp",
                    "type": "string"
                },
                "website": {
                    "description": "Organization website URL",
                    "type": "string",
                    "example": "https://www.wecareholidays.com"
                }
            }
        },
...
        "models.SwaggerErrorResponse": {
            "description": "Error response model",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Optional error details"
                },
                "errorMessage": {
                    "description": "Human-readable error message",
                    "type": "string",
                    "example": "Invalid request parameters"
                },
                "statusCode": {
                    "description": "HTTP status code",
                    "type": "integer",
                    "example": 400
                },
                "success": {
                    "description": "Indicates request failure",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.SwaggerStandardResponse": {
            "description": "Standard API response format",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response payload"
                },
                "errorMessage": {
                    "description": "Error message (only present if success is false)",
                    "type": "string",
                    "example": ""
                },
                "statusCode": {
                    "description": "HTTP status code",
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "description": "Indicates whether the request was successful",
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}